refactor(domain) : CrewAttendanceRecords 생성 방식 리팩터링
- 팩토리 메서드 패턴 사용
- CsvParsingGenerator 추가
