+++ src/main/java/domain/CrewAttendanceRecordsGenerator.java
@@ -1,0 +1,7 @@
+package domain;
+
+import java.util.Map;
+
+public interface CrewAttendanceRecordsGenerator {
+    public Map<Crew, AttendanceRecords> generate(DateGenerator dateGenerator);
+}
