+++ src/test/java/domain/AttendanceRecordsTest.java
@@ -13,7 +13,7 @@ class AttendanceRecordsTest {
     @Test
     @DisplayName("출석 기록이 없는 날짜가 결석으로 기록되었는지 확인한다.")
     void fillAbsencesTest() {
-        CrewAttendanceRecords crewAttendanceRecords = new CrewAttendanceRecords("/attendances.csv", () -> LocalDate.of(2024, 12, 13));
+        CrewAttendanceRecords crewAttendanceRecords = new CrewAttendanceRecords(new CsvParsingGenerator(), () -> LocalDate.of(2024, 12, 13));
         Crew crew = new Crew("쿠키");
         boolean hasRecord = crewAttendanceRecords.hasRecord(crew, LocalDate.of(2024, 12, 12));
